refactor: Arrows.tsx 위치 변경
